requirements.txt
fastapi
uvicorn
python-multipart
PyPDF
transformers
pymilvus
pydantic
pyyaml
sqlalchemy
dotenv
PyPDF2
python-multipart
pdfminer.six
sqlalchemy
alembic
psycopg2-binary